feat(expo): Only bundle Clerk in native platforms
