+++ packages/expo/src/provider/singleton/index.ts
@@ -1,0 +1,1 @@
+export * from './singleton';
