+++ packages/expo/src/provider/singleton/singleton.ts
@@ -1,0 +1,9 @@
+import { Clerk } from '@clerk/clerk-js/headless';
+
+import { createClerkInstance } from './createClerkInstance';
+
+/**
+ * @deprecated Use `getClerkInstance` instead. `Clerk` will be removed in the next major version.
+ */
+export { clerk } from './createClerkInstance';
+export const getClerkInstance = createClerkInstance(Clerk);
